backend/yourmirror/settings/prod.py
import os
from .base import PROJECT_ROOT, env
import environ


environ.Env.read_env(os.path.join(PROJECT_ROOT, ".env"))

DEBUG = False

EMAIL_BACKEND = "django.core.mail.backends.smtp.EmailBackend"
EMAIL_PORT = 465
EMAIL_USE_SSL = True
EMAIL_HOST = env("EMAIL_HOST", default="")
EMAIL_HOST_USER = env("EMAIL_HOST_USER", default="")
EMAIL_HOST_PASSWORD = env("EMAIL_HOST_PASSWORD", default="")
DEFAULT_FROM_EMAIL = env("DEFAULT_FROM_EMAIL", default="")
