backend/pyproject.toml
[project]
name = "yourmirror.backend"
version = "1.0.0"
dependencies = [
    "django-tailwind>=4.0.1",
    "django>=5.2",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "django-environ>=0.12.0",
    "django-minio-storage>=0.5.7",
    "celery[redis]>=5.5.2",
]

[dependency-groups]
test = [
    "factory-boy>=3.3.3",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.10.0",
]
dev = [
    "cookiecutter>=2.6.0",
    "django-tailwind[reload]>=4.0.1",
]
lint = [
    "django-stubs[compatible-mypy]>=5.2.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "yourmirror.settings.test"
python_files = ["test_*.py"]
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*",
]

[tool.pyright]
ignore = [
  "tests/factories.py"
]

[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR"
plugins = ["mypy_django_plugin.main"]

show_error_codes = true
strict = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "environ",
    "celery",
]
ignore_missing_imports = true

# для игнора вызова всех фабрик
[[tool.mypy.overrides]]
module = [
    "tests.*",
]
disable_error_code = ["no-untyped-call"]

[[tool.mypy.overrides]]
module = [
    "tests.factories",
    "manage",
]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "yourmirror.settings.test"
